Move new module to its own file.
