+++ rust/src/checkpoints.rs
@@ -17,9 +17,9 @@ use super::delta_arrow::delta_log_schema_for_table;
 use super::open_table_with_version;
 use super::schema::*;
 use super::storage::{StorageBackend, StorageError};
+use super::table_state::DeltaTableState;
 use super::writer::time_utils;
 use super::{CheckPoint, DeltaTableError};
-use crate::delta::table_state::DeltaTableState;
 use crate::DeltaTable;
 
 /// Error returned when there is an error during creating a checkpoint.
